requirements.txt
# Core Telegram Bot dependencies
python-telegram-bot==20.3
aiosqlite==0.19.0

# Flask API dependencies
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1

# Web server for health checks and keep-alive
aiohttp==3.9.1
requests==2.31.0

# Environment and configuration
python-dotenv==1.0.0

# Async utilities
nest-asyncio==1.5.8

# Additional utilities for production
python-dateutil==2.8.2
